Add files via upload
The Python Parking Lot project allows a user to enter their vehicle’s type and license plate number. The system shows a parking grid that allows the user to see the number of available spaces as well as what types of vehicles are currently parked within the lot. The system also tracks a user’s time within the lot and will calculate a fare when the user checks out their vehicle.