Add files via upload
The Zodiac Sign Calculator In Python is a simple project developed using Python. The project is for the user’s convenience, to help them find out about their zodiac sign with the help of birthdate and month. 
This is a simple GUI-based project which is very easy to understand and use. Also, this project makes a convenient way for the user to discover their zodiac sign.
This simple project is in Python. Talking about the features of this system, this python application is for calculating the zodiac sign of the users. Here, you can simply enter the birth date, and month in the text fields and click the ‘Get Zodiac Sign’ button to find your zodiac symbol.