Add files via upload
The Simple Age Calculator In Python is a simple project developed using Python. The project is for the user’s convenience, to help them find out about their exact age in year, month, and days. This is a simple GUI-based project which is very easy to understand and use. Also, this project makes a convenient way for the user to discover the exact age.

This simple project is in Python. Talking about the features of this system, this python application is for calculating the age of the users. So, you can simply enter the birth date, and given date in day, month, and year in the text fields and click the ‘Get Age’ button.