Add files via upload
The Sudoku Solver In Python is a simple project developed using Python. This is a simple GUI-base project which is very easy to understand and use. This game is a simple puzzle project. You have to input numbers from 1 to 9 in order to complete the puzzle box.

Talking about the features of this system, this python application is designed to solve number puzzles. You can use solver from the option to solve the puzzle if you want. Also, you can submit your answer after completing the puzzle. It is also capable of handling all types of exceptions. Also, the design of this system is pretty simple so that the user won’t get any difficulties while working on it.