Add files via upload
The Kids Mind Trainer In Python is a simple project for helping your kids grow in IQ. The project contains only the user side. The user can start the game by clicking on the ok button. Also, you can see some images on the screen and choose the right answer according to the images. The user can refresh the image items if they do not know the correct word for it.

It is a learning game based on Python Gui. This app project just contains the user section. The user has to click on the ok button, then choose the words that match with the images. The different images will be displayed on the screen and you have to select its correct word. You can change the images if you are not able to guess the correct form. The design of this project is pretty simple so that the user won’t find any difficulties while working on it.