Add files via upload
The Typing Speed In Python is a simple project for improving your Typing speed. The project contains only the user side. It is a simple typing challenge game mini-project. This is like a game, one-by-one random word will pop up on your screen and you have to type as many given words as you can, within 60 seconds.

This app project just contains the user section. The player/user can improve their Typing speed playing this game. You have to type as many words as you can in the given amount of seconds to score. Otherwise, you will lose the game, and to play again, you just have to type the current word. Also, your score will be reset and the high score will be set on the game. The design of this project is pretty simple so that the user won’t find any difficulties while working on it.