Add files via upload
The Binary converter In Python is a simple project developed using Python. The project contains a base conversion system. The user can enter any number and convert them to other number systems (decimal to binary, binary to decimal).
This is a simple command-line based project which is very easy to understand and use. Also, this project makes a convenient way for the user to gain an idea of how to perform number conversion.