Add files via upload
The Color Game In Python is a simple project developed using Python. This project is a GUI based which displays random color text and you have to identify its color. You have 30 seconds to guess as much as word color you can, and you will get score if the guess is right. This project is an interesting and simple project.

This Color Game application is designed for guessing random word color. Here, you have to type the color of the words, not the word text. Also, the design of this system is pretty simple so that the user won’t get any difficulties while working on it.