Add files via upload
This python game application is designed to play and solve the cube using arrows for the movements. Module Used – Tkinter()-It is a standard Python interface to the Tk GUI toolkit shipped with Python. Python with tkinter outputs the fastest and easiest way to create the GUI applications. Creating a GUI using tkinter is an easy task. Also, the design of this system is pretty simple so that the user won’t get any difficulties while working on it.