Add files via upload
The Rocket Game project contains few lines of codes in Python. The gameplay is GUI based and the controls are simple for the users. This is one of the most addictive games. The is similar to flappy bird game.
The playing methods are the same as that of the flappy bird, all you have to do is click on the rocket continuously to move it upward. If the player stops to click on the rocket, then it will fall down and you will lose. The gameplay design is simple that the user won’t find it difficult to use and navigate.