Add files via upload
Simple 2048 Game project is written in Python. The gameplay is GUI based and the controls are simple for the users. This is one of the most addictive games. All the playing methods are the same, all you have to do is move numbers on a grid to merge them to form the number 2048; after all, you can keep on playing the game by creating larger numbers.
In each move, a new number of displays (2 or 4). The player has to move the numbers using arrow keys and try to collide the same two numbers which create a new number with the total sum of the two numbers. The gameplay design is simple that the user won’t find it difficult to use and navigate.