Add files via upload
The Simple Air Hockey Game project is written in Python. It’s a two-player game, playing against each other. After starting the game, the player has to move the paddle from left to right, up and down to block the ball from entering the scoring area. And, the game consists of 7 rounds and the player who wins two/or more rounds will be the winner.

A simple and clean GUI is provided for easy gameplay. Talking about the features of the game, the player has to move the paddle from left to right, up and down to block the ball. It’s a two-player game, playing against each other. One user can use the arrow keys and the next can use WASD keys for the movement of the ball. The gameplay design is so simple that the user won’t find it difficult to use and understand.