Add files via upload
Catch the fruit in the python is a simple game.  GUI uses the pygame library. Talking about the gameplay, it’s a simple single player platformer game, where the player has to collect all the candies in order to win the game. The main objective of this game is to take as many fruits as you can. This game is inspired by the 90s famous game “fruit-ninja”.
A simple and clean GUI is provided for easy gameplay. The gameplay design is so simple that the user won’t find it difficult to use and understand. Different images are used in the development of this game project, the gaming environment is just like the fruit ninja game.