Add files via upload
Snakes and Ladders Game project is written in Python. This is a GUI based board game which is very easy to understand and use. Talking about the gameplay it’s all same as the real one. The user can select whether to play multiplayer or with the computer.
After starting the game, the player has to roll the dice and after rolling it the game moves the token automatically according to the dice number. The gameplay is similar to the real one. Here, the player also gets another chance to roll the dice whenever he/she gets 6 number. There are numbers of ladders and snakes in the game which helps the player to upgrade or downgrade the square number. The player who reaches the last square of the track is the winner.
This is a 2D GUI Based game written for the beginners. Snakes and Ladders Game in Python project with source code is free to download. Use for education purpose only! For the project demo, have a look at the video below.