Add files via upload
Simple Spider-man Game project is written in Python.  This is a simple GUI based action game which is very easy to understand and use. Talking about the gameplay, the user just has to fight with the villain of Spiderman and win the battle in order to proceed towards the next stage.
The game is made up of different scenes.
After starting the game, the player (Spiderman) has to go nearer to the villain in order to start the battle. The battle strategy is too simple, it’s like a word attack game. The player just has to hit “A” key to attack the villain. The player and villain have the health of 30 bars. After winning the battle, the player can proceed towards the next level. Talking about gaming environment, the project uses various images and GIF files to provide real-time experience.
This is a simple 2D GUI Based game written for the beginners. Simple Spider-man Game in Python project with source code is free to download. Use for education purpose only! For the project demo, have a look at the image slider below.