Add files via upload
Simple 2D Ludo Game project is written in Python. This is a simple GUI based strategy board game that is very easy to understand and use. Talking about the gameplay, all the playing rules are the same just like we play in real-time ludo. This is a simple 2D multiplayer game.
After starting the game, a GUI ludo board appears, other rules are the same. First, the player has to roll the dice. The main thing in this simple GUI based game is that the player just has to press “Roll” to roll the dice. At the top of the board, it displays a dice with the number. The has to keep on rolling until there’s a possible pawn to move. All the game movements are to be performed manually by the player. A simple 2D GUI is provided for easy gameplay. The gameplay design is so simple that the user won’t find it difficult to use and understand. For the gaming environment, the project uses various images and GIF files to provide real-time experience.