Add files via upload
F1 Race Road Game project is a python program and sound files. The gameplay Graphics are pretty good and the controls are too simple for the users. 

Talking about the gameplay, all the playing methods are too simple all you have to do is just dodge the f1 from the opponent. Here, the user has to dodge the f1 from other opponents in order to gain score points. 

The more players keep on playing, more will be the increment in the gameplay speed. Whenever the player touches the opponents or gets to the end of the screen, the game is over and the user can check for his/her score in the top left corner with the number of opponents dodged. The main objective of this game is to score points as much as possible. A simple GUI is provided for easy gameplay. The gameplay design is so simple that the user won’t find it difficult to use and navigate.