Add files via upload
Tanks Game project gameplay is GUI based and the controls are simple for the users. Talking about the gameplay, it’s one of the most addictive game. All the playing methods are simple, the main objective of this game is to shoot and destroy the opponent’s tank before getting destroyed. Enemies get harder when the player shoots more and more. 
The player has to move the tank using left-right arrow keys and try to destroy the opponent tanks using Space Bar. Remember there’s a Power bar at the top, and the player has to maintain power level too in order to attack.
Otherwise, the player fails to attack and the bullets fired cannot cross the wall. The wall is created with random heights on each game. So the player has to maintain the power level to cross the fired bullets across the wall in order to destroy the opponent. The player has to move the turret upward or downward according to the wall’s height by using Up and Down arrow keys. From the terminal, the user can check the player and enemies damage after the attack. The gameplay design is too simple that user won’t find it difficult to use and navigate.