Add files via upload
Aircraft War Game project is a python gameplay. Graphics is good enough and the controls are simple for the users. 
Talking about the gameplay, it’s a common game for all. All the playing methods are too simple all you have to do is just move around and destroy enemies by firing bullets. Here, the user has to control the aircraft using up, down, left, right arrow keys and the game score depends upon the player’s number of kills. 
This means the more you destroy enemies, more will be the game score. The player loses whenever the enemies come close to the player. A simple GUI with sound is provided for the easy gameplay. The gameplay design is simple that user won’t find it difficult to use and navigate.